refactor: 오류 처리 및 코드 구조 개선, 비동기 요청 업데이트
- AdminService 메소드 수정: try-catch 추가해 delete, toggle, report 처리 시 CustomException 반환 로직 구현
- AdminRepository 메소드의 매개변수 타입 int → Long으로 변경
- SecurityConfig 로그인 실패 시 DisabledException 처리 및 동적 리다이렉션 로직 추가
- 불필요한 AdminVO 클래스 삭제

+++ src/main/java/com/travel/letsgospringboot/admin/repository/AdminRepository.java
@@ -11,20 +11,30 @@ import java.util.List;
 @Mapper
 public interface AdminRepository {
     int selectTotalUserCount();
+
     int selectTotalPlaceCount();
+
     int selectTotalScheduleCount();
+
     List<PlaceVO> selectRecentPlaces();
 
     List<PlaceVO> selectAdminPlaces(@Param("keyword") String keyword);
+
     int insertPlace(PlaceVO placeVO);
+
     int updatePlace(PlaceVO placeVO);
-    int deletePlace(@Param("placeId") int placeId);
-    int updatePlaceVisibility(@Param("placeId") int placeId, @Param("isActive") boolean isActive);
+
+    int deletePlace(@Param("placeId") Long placeId);
+
+    int updatePlaceVisibility(@Param("placeId") Long placeId, @Param("isActive") boolean isActive);
 
     List<AdminReportVO> selectAllReports();
+
     int updateReportStatus(@Param("reportId") Long reportId, @Param("status") String status);
+
     AdminReportVO selectReportById(@Param("reportId") Long reportId);
 
     List<AdminPostVO> selectAdminPosts(@Param("keyword") String keyword);
+
     int updatePostVisibility(@Param("postId") String postId, @Param("isHidden") int isHidden);
 }
